order more fully implemented in buyer

+++ deliverer/delete_deliverer.php
@@ -8,6 +8,19 @@
 		die( "Connection failed: ".$conn->connect_error);
 	}
 	
+	$select = "SELECT * FROM DELIVEREDBY WHERE deliverer_username=? AND is_delivered=0";
+	$stmt = $conn->prepare($select);
+	$stmt->bind_param("s", $username);
+	$stmt->execute();
+	$res = $stmt->get_result();
+	if ($res->row_nums != 0) {
+		$conn->close();
+		header( "Location: deliverer.php?error=Cannot%20Delete%20Account%20With%20Outstanidng%20Deliveries" );
+		exit ;
+	}
+	echo $res->row_nums;
+	
+	/*
 	$insert = "DELETE FROM MANAGES WHERE username=?;";
 	$stmt = $conn->prepare($insert);
 	$stmt->bind_param("s", $username);
@@ -24,5 +37,6 @@
 
 	header( "Location: ../index.php?error=Successfully%20Deleted%20Account&logout=true" );
 	exit ;
+	*/
 ?>
 </html>
